chore: Move Error to a new package
